Few opcodes Thanks to StrawberryEmu for SMSG_LEARNED_SPELL

+++ src/server/game/Entities/Player/Player.cpp
@@ -3990,7 +3990,7 @@ void Player::learnSpell(uint32 spell_id, bool dependent)
     {
         WorldPacket data(SMSG_LEARNED_SPELL, 6);
         data << uint32(spell_id);
-        data << uint16(0);
+        data << uint32(0);
         GetSession()->SendPacket(&data);
     }
